requirements.txt
requests
beautifulsoup4
urllib3
dnspython
whois
colorama
